cgo: Keep alive.
When testing on android within cake, libwallet is being garbage
collected by the os when in the background. These changes seem to keep
that from happening but may not currenlty be optimal. Also changes the
wallet map to only allowing one at a time as that is how we are using
it.